+++ examples/map_text_progressively.py
@@ -14,7 +14,7 @@ second_upload = documents[500:]
 project = atlas.map_text(data=first_upload,
                           indexed_field='text',
                           name='News 1k Example Progressive',
-                          num_workers=10)
+                        )
 
 print(project.maps)