+++ docs/release_notes.md
@@ -2,7 +2,7 @@
 ## v1.0.25
 **Tagging**: The `get_tags` method will retrieve tags you have assigned to datapoints on the map.
 
-**Progressive projects**: You can now call the `map_*` endpoints multiple times and specify the same project each time. Doing this will add data to the project. See the [documentation](updating_maps.md) for examples.
+**Progressive projects**: You can now call the `map_*` endpoints multiple times and specify the same project each time. Doing this will add data to the project. See the [documentation](dynamic_maps.md) for examples.
 
 **shard_size**: You can now specify a shard_size in the `map_*` endpoints. If each datum is too large, you want to use a smaller shard size to successfully send data to Atlas.